Fix ambiguity about whether final bit padding must be zero
* Currently, the wording makes it clear that fill bits must be zero if
ISLASTEMPTY is true. However, it says nothing about whether the fill bits must
be zero if the meta-block is the last, but is also non-empty. The behavior of
the reference implementation is that it must be zero. Thus, move the requirement
of fill bits being zero to the ISLAST description since that covers ISLASTEMPTY.

* Also, make the description of ISUNCOMPRESSED consistent with the way meta
data are described.