appveyor: also upload source distribution packages to Github Releases
This is so that one can do:

    pip install --find-links https://github.com/google/brotli/releases/latest brotli

on any platform, whether or not there are pre-compiled *.whl packages that match the user's python version and anchitecture (e.g. Linux). Pip automatically chooses the correct binary package, and if that isn't available the source distribution (*.tar.gz or *.zip) is downloaded.